ScoutAgent/pyproject.toml
[project]
name = "scout-agent"
version = "0.1.0"
description = "Broker tools agent using AWS Bedrock AgentCore Gateway with Strands SDK"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.41.0",
    "strands-agents>=1.18.0",
    "strands-agents-tools>=0.2.0",
    "mcp>=1.0.0",
    "httpx-auth-awssigv4>=0.1.0",
    "requests>=2.31.0",
    "requests-aws4auth>=1.2.0",
    "bedrock-agentcore>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
